+++ src/Vankosoft/CmsBundle/Controller/VankosoftFileManagerController.php
@@ -3,7 +3,7 @@
 use Symfony\Component\HttpFoundation\Request;
 
 use Vankosoft\ApplicationBundle\Controller\AbstractCrudController;
-use Vankosoft\ApplicationBundle\Controller\TaxonomyHelperTrait;
+use Vankosoft\ApplicationBundle\Controller\Traits\TaxonomyHelperTrait;
 
 class VankosoftFileManagerController extends AbstractCrudController
 {