+++ src/Vankosoft/CmsBundle/Form/PageCategoryForm.php
@@ -3,6 +3,7 @@
 use Vankosoft\ApplicationBundle\Form\AbstractForm;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
@@ -73,6 +74,17 @@ class PageCategoryForm extends AbstractForm
                 
                 'required'      => false,
                 'placeholder'   => 'vs_cms.form.category.parent_category_placeholder',
+            ])
+        ;
+    }
+    
+    public function configureOptions( OptionsResolver $resolver ): void
+    {
+        parent::configureOptions( $resolver );
+        
+        $resolver
+            ->setDefaults([
+                'csrf_protection' => false,
             ])
         ;
     }